move community to legacy
