+++ localstack-core/localstack/services/apigateway/legacy/context.py
