+++ localstack-core/localstack/services/apigateway/legacy/integration.py
@@ -22,7 +22,6 @@ from localstack.aws.connect import (
 )
 from localstack.constants import APPLICATION_JSON, HEADER_CONTENT_TYPE
 from localstack.services.apigateway import helpers
-from localstack.services.apigateway.context import ApiInvocationContext
 from localstack.services.apigateway.helpers import (
     ApiGatewayIntegrationError,
     IntegrationParameters,
@@ -34,7 +33,8 @@ from localstack.services.apigateway.helpers import (
     make_error_response,
     multi_value_dict_for_list,
 )
-from localstack.services.apigateway.templates import (
+from localstack.services.apigateway.legacy.context import ApiInvocationContext
+from localstack.services.apigateway.legacy.templates import (
     MappingTemplates,
     RequestTemplates,
     ResponseTemplates,
