+++ tests/unit/test_apigateway.py
@@ -25,21 +25,21 @@ from localstack.services.apigateway.helpers import (
     extract_query_string_params,
     get_resource_for_path,
 )
-from localstack.services.apigateway.integration import (
+from localstack.services.apigateway.legacy.integration import (
     LambdaProxyIntegration,
     apply_request_parameters,
 )
-from localstack.services.apigateway.invocations import (
+from localstack.services.apigateway.legacy.invocations import (
     ApiInvocationContext,
     BadRequestBody,
     RequestValidator,
 )
-from localstack.services.apigateway.models import ApiGatewayStore, RestApiContainer
-from localstack.services.apigateway.templates import (
+from localstack.services.apigateway.legacy.templates import (
     RequestTemplates,
     ResponseTemplates,
     VelocityUtilApiGateway,
 )
+from localstack.services.apigateway.models import ApiGatewayStore, RestApiContainer
 from localstack.testing.config import TEST_AWS_REGION_NAME
 from localstack.utils.aws.aws_responses import requests_response
 from localstack.utils.common import clone
